requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-auth==2.27.0
selenium==4.0.0
webdriver-manager==3.8.0
python-dotenv==1.0.0




